+++ docs/release-notes.md
@@ -1,3 +1,8 @@
+## 5.1.1
+
+- Option to show toast notification when BT handles links.
+- If any issues are detected, health icon will flash and health dashboard will stay open as a window. Previously you had an annoying popup opening every minute or so.
+
 ## 5.1.0
 
 ### New features